refactor: RealWorkCode 클래스 component 패키지로 이동
