+++ src/main/java/project/volunteer/domain/storage/application/StorageService.java
@@ -1,7 +1,7 @@
 package project.volunteer.domain.storage.application;
 
 import org.springframework.web.multipart.MultipartFile;
-import project.volunteer.domain.image.domain.RealWorkCode;
+import project.volunteer.global.common.component.RealWorkCode;
 import project.volunteer.domain.storage.domain.Storage;
 
 public interface StorageService {
