+++ src/main/java/project/volunteer/domain/user/application/UserServiceImpl.java
@@ -1,19 +1,10 @@
 package project.volunteer.domain.user.application;
 
-import java.util.Optional;
-
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import project.volunteer.domain.image.application.ImageService;
-import project.volunteer.domain.image.application.dto.ImageParam;
-import project.volunteer.domain.image.dao.ImageRepository;
-import project.volunteer.domain.image.domain.Image;
-import project.volunteer.domain.image.domain.ImageType;
-import project.volunteer.domain.image.domain.RealWorkCode;
 import project.volunteer.domain.user.dao.UserRepository;
 import project.volunteer.domain.user.domain.User;
 import project.volunteer.global.error.exception.BusinessException;
