+++ src/main/java/project/volunteer/domain/user/dao/queryDto/UserQueryDtoRepositoryImpl.java
@@ -15,7 +15,7 @@ import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import lombok.RequiredArgsConstructor;
-import project.volunteer.domain.image.domain.RealWorkCode;
+import project.volunteer.global.common.component.RealWorkCode;
 //import project.volunteer.domain.user.api.dto.HistoryTimeInfo;
 import project.volunteer.domain.user.dao.queryDto.dto.QUserRecruitingQuery;
 import project.volunteer.domain.user.dao.queryDto.dto.QUserRecruitmentJoinRequestQuery;
